requirements.txt
fastapi
uvicorn
SQLAlchemy
pytest
httpx
passlib[bcrypt]
python-jose[cryptography]
pydantic-settings
python-multipart
